feat(components/editabletext): create component (#1579)
