+++ packages/components/src/EditableText/index.js
@@ -1,0 +1,3 @@
+import EditableText from './EditableText.component';
+
+export default EditableText;
