--- packages/components/src/SubHeaderBar/InputTitleSubHeader/InputTitleSubHeader.scss
@@ -1,140 +1,0 @@
-@import '~@talend/bootstrap-theme/src/theme/animation';
-
-@mixin ellipsis {
-	white-space: nowrap;
-	text-overflow: ellipsis;
-	overflow: hidden;
-}
-
-@mixin input-text($weight) {
-	color: $black;
-	font-size: $tc-input-subheader-size-medium;
-	font-weight: $weight;
-}
-
-$tc-input-subheader-size-large: 2.4rem !default;
-$tc-input-subheader-size-medium: 1.4rem !default;
-$tc-input-subheader-size-small: 1rem !default;
-$tc-input-subheader-title-weight: 900 !default;
-$tc-input-subheader-input-weight: $form-base-font-weight !default;
-$tc-input-subheader-sub-title-weight: 300 !default;
-
-:global(.tc-subheader-details-blink) {
-	@include heartbeat(object-blink);
-}
-
-.tc-subheader-details {
-	display: flex;
-	margin: 0;
-	height: 100%;
-
-	&-icon {
-		align-self: center;
-		width: $tc-input-subheader-size-large;
-		height: $tc-input-subheader-size-large;
-		color: $brand-primary;
-		margin-right: $padding-large;
-	}
-
-	& &-text {
-		display: flex;
-		flex-direction: column;
-		justify-content: center;
-		overflow: hidden;
-		flex: 1 auto;
-		margin-right: 110px;
-		max-width: 90rem;
-
-		&-title {
-			display: inline-flex;
-			align-items: center;
-
-			&-wording,
-			&-wording-button {
-				@include input-text($tc-input-subheader-title-weight);
-				@include ellipsis;
-				text-transform: uppercase;
-				margin: 0;
-			}
-
-			&-wording-button {
-				padding: 0;
-				letter-spacing: inherit;
-
-				&:global(.btn-link) {
-					color: $black;
-
-					&:focus,
-					&:hover,
-					&:active {
-						color: $black;
-					}
-				}
-			}
-
-			&-pencil {
-				padding-left: $padding-small;
-				padding-right: $padding-small;
-
-				&:global(.btn-link) {
-					color: $silver;
-				}
-			}
-		}
-
-		&-subtitle {
-			@include ellipsis;
-			color: $dark-silver;
-			font-size: $tc-input-subheader-size-medium;
-			font-weight: $tc-input-subheader-sub-title-weight;
-		}
-	}
-	& &-form {
-		display: flex;
-		align-items: center;
-		padding-left: $padding-large;
-
-		& .tc-subheader-details-form-input {
-			@include input-text($tc-input-subheader-input-weight);
-			width: 18rem;
-			&::selection {
-				background: lighten($scooter, 25%);
-			}
-		}
-
-		&-buttons {
-			display: flex;
-
-			&-icon {
-				border: 0.2rem solid currentColor;
-				width: $tc-input-subheader-size-large;
-				height: $tc-input-subheader-size-large;
-				min-height: auto;
-				display: flex;
-				border-radius: 50%;
-				padding: 0;
-				margin-left: $padding-large;
-				justify-content: center;
-				align-items: center;
-
-				> svg {
-					width: $tc-input-subheader-size-medium;
-					height: $tc-input-subheader-size-medium;
-				}
-
-				&:last-child {
-					margin-left: $padding-small;
-
-					> svg {
-						width: $tc-input-subheader-size-small;
-						height: $tc-input-subheader-size-small;
-					}
-				}
-			}
-		}
-
-		:global(.form-group) {
-			padding-top: 0;
-		}
-	}
-}
